requirements.txt
streamlit==1.31.0
pandas==2.1.1
matplotlib==3.8.0
python-dotenv==1.0.1
requests==2.31.0
boto3>=1.28.57
biopython
langchain-community==0.3.1

